Cast to float32.

+++ python/train_ghostprob.py
@@ -178,7 +178,7 @@ if arguments.plot:
 if arguments.int8:
     print("INT8 quantization")
     def representative_data_gen():
-        for input_value in tf.data.Dataset.from_tensor_slices(data_test).batch(1).take(100):
+        for input_value in tf.data.Dataset.from_tensor_slices(data_test.astype("float32")).batch(1).take(100):
             yield [input_value]
     converter = tf.lite.TFLiteConverter.from_keras_model(model)
     converter.optimizations = [tf.lite.Optimize.DEFAULT]
